+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+### OCD-1449
+* Add notifications service to allow registration, getting, and deleting of email subscriptions
+* Add /data/notification_types service to send back the types of notifications a logged-in user may work with
+
+---
+
 ## Version TBD
 _Date TBD_